+++ ccp4i2/__init__.py
@@ -1,3 +1,3 @@
-__version_date__ = "31 Jan 2025 14:20"
-__version_info__ = (2, 2, 2)
+__version_date__ = "7 Feb 2025 11:00"
+__version_info__ = (2, 2, 3)
 __version__ = ".".join(map(str, __version_info__))